# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pulseboard API
  version: 1.1.0
  description: v1.1 API for authentication, current-user access, and owner-only boards.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Process health
      responses: { '200': { description: Service is running } }
  /ready:
    get:
      summary: Database readiness
      responses: { '200': { description: Database is reachable }, '503': { description: Database unavailable } }
  /api/auth/register:
    post:
      summary: Register a user
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AuthRequest' } } } }
      responses: { '201': { description: Registered }, '400': { $ref: '#/components/responses/InvalidRequest' }, '409': { description: Email already exists }, '429': { $ref: '#/components/responses/RateLimited' } }
  /api/auth/login:
    post:
      summary: Login and receive a JWT
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AuthRequest' } } } }
      responses: { '200': { description: Token issued, content: { application/json: { schema: { type: object, required: [token], properties: { token: { type: string } } } } } }, '400': { $ref: '#/components/responses/InvalidRequest' }, '401': { description: Invalid credentials }, '429': { $ref: '#/components/responses/RateLimited' } }
  /api/me:
    get:
      summary: Current authenticated user
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: User profile, content: { application/json: { schema: { $ref: '#/components/schemas/UserProfile' } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/boards:
    post:
      summary: Create a board
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateBoardRequest' }, example: { title: Launch Plan, description: v1.1 rollout board } } } }
      responses: { '201': { description: Board created, content: { application/json: { schema: { $ref: '#/components/schemas/Board' } } } }, '400': { $ref: '#/components/responses/InvalidRequest' }, '401': { $ref: '#/components/responses/Unauthorized' } }
    get:
      summary: List current user's boards
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Boards ordered newest first, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Board' } } } } }, '401': { $ref: '#/components/responses/Unauthorized' } }
  /api/boards/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
    get:
      summary: Get an owned board
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Board, content: { application/json: { schema: { $ref: '#/components/schemas/Board' } } } }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/BoardNotFound' } }
    patch:
      summary: Update an owned board
      security: [{ bearerAuth: [] }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/UpdateBoardRequest' }, example: { title: Updated Launch Plan, description: Revised rollout board } } } }
      responses: { '200': { description: Updated board, content: { application/json: { schema: { $ref: '#/components/schemas/Board' } } } }, '400': { $ref: '#/components/responses/InvalidRequest' }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/BoardNotFound' } }
    delete:
      summary: Soft-delete an owned board
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Board deleted }, '401': { $ref: '#/components/responses/Unauthorized' }, '404': { $ref: '#/components/responses/BoardNotFound' } }
components:
  securitySchemes: { bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
  schemas:
    AuthRequest: { type: object, required: [email, password], properties: { email: { type: string, format: email }, password: { type: string, minLength: 6 } } }
    UserProfile: { type: object, required: [id, email, created_at], properties: { id: { type: integer }, email: { type: string, format: email }, created_at: { type: string, format: date-time } } }
    CreateBoardRequest: { type: object, required: [title], properties: { title: { type: string, minLength: 1 }, description: { type: string, nullable: true } } }
    UpdateBoardRequest: { type: object, properties: { title: { type: string, minLength: 1 }, description: { type: string, nullable: true } } }
    Board: { type: object, required: [id, owner_id, title, created_at, updated_at], properties: { id: { type: integer }, owner_id: { type: integer }, title: { type: string }, description: { type: string, nullable: true }, created_at: { type: string, format: date-time }, updated_at: { type: string, format: date-time } } }
  responses:
    InvalidRequest: { description: Invalid request body or fields }
    Unauthorized: { description: Missing, invalid, expired, or incorrectly signed token }
    BoardNotFound: { description: Board does not exist, is deleted, or belongs to another user }
    RateLimited: { description: Too many requests; inspect Retry-After header }